Fixed "send message" window.
Formerly there was inconsistency, with console private messages limited to 350 characters, but "send message" from Nations tab limited to 80 characters. This discouraged use of the latter feature. 80 characters is not enough for any kind of negotiation or diplomacy.  This push can't be tested locally so is not guaranteed to work right. This is a test.