Update to Cities List
Continued Rhue's great work on this and copied some of his code and ideas to evolve this even further.  All columns are sortable, FPT and GLS have icon graphics now.  And shows mini-sprite for what is being produced.  In addition, used alignment/spacing/boldness/colour/blank space omission, as tools to increase easier readability of all info in a glance.