Withdraw Vision button, dirty hack fix
Ideally the button should be disabled when there is no vision to withdraw. However, the former code disabled a delegate control player from withdrawing vision. Having an enabled button that does nothing is a better condition than a disabled button in cases where it needs to be used.  This is a hack fix and hopefully someone will rewrite this to properly disable the button again but not for someone controlling the delegate.  Also this is a test to see if it works.