Sync.sh defaults to skip flags
Flags never changed but made the script take 20x longer to complete.

In addition, the script was needlessly overwriting flags that we fixed, because of faulty svg>png conversion on about 100 different flags.

The old version with flags is sync-with-flags.sh.  It can be used if/when new flags are ever added.  If so, care should be taken to backup and overwrite the 100 or so fixed flags that svg conversion was breaking.