Reverted unclickable containers for status and orders buttons
But left them as comments so we can try to fix it again in future.
It succeeded in not making the container block map clicks, but the children().css(pointer-events:auto) was not making contained elements override to be clickable like expected. Needs further investigation or another solution like not setting width to be whole screen wide.