Change min width for status bar to go at top.
It was going to the bottom even when it had enough room on top. 745ish px was the measured space it had on top (which was adequate), yet it was requiring >800 and going to the bottom status bar instead, which we would prefer not to do since that item prevents map clicks.