mp2sandbox default settings for traderoutes
Simple trade revenue style: no distance multipliers or OP multipliers of any kind. Allows us to tune it just with normal multipliers/dividers in game.ruleset, in order to make sure trade routes are a MINOR yet VIABLE bonus: i.e. unnecessary and a mre option in a larger menu of things giving ROI. mp2sandbox is experimenting with a middle path between "trade routes are good and add more to the game and make it more than just war" and "trade routes are bad and a dominating OP element of the game that consumes massive micromanagement and minimises all the other elements."  MP2 will err on the side of less complication,micro-management,and avoiding an OP element--leaning more true to the decision original MP made by eliminating traderoutes totally. It may keep them out altogether, this is only a sandbox experiment after all