temp fix, mapview_mouse_movement mode was preventing clicking on cities
Because it gets set to on when it's not REALLY on yet. TO DO: make another var for REALLY knowing if we're in map-drag, that goes ON the first time it actually starts dragging the map and changing the cursor to mapdrag mode. Cursor changing to map-drag could serve for that.