Insert logic for "true" goto_path.length to handle fuel units
Units with fuel such as Trireme, Fighter, report a goto_path.length +1 higher than it really is. This was interfering with logic to override GOTO command when given to an adjacent tile, in order to allow it to be possible in cases the server uses "nanny mode" to disallow the GOTO / force it to wait until next turn.