GOTO path will redraw several times on the same tile before blocking requests for new paths on the same tile
The performance optimisation to not ask for new paths from the server constantly on the same tile, was much faster but occasionally was "too fast" and would draw a path before it had time to construct it then block further updates on paths to the same tile, not allowing it to finally construct it right. This now allows tuning it to request new paths x times before blocking further requests.