Add SafeLink functionality
In-game links with %%% instead of %% will override to always go to oneself as a message.  This facilitates out-of-game pasting of %%%links to each other and the recipient to just paste it in game to be able to use the link; instead of awkwardly having to send it to someone else or oneself.