Fix action_max_range load error handling.
The switch() for entry types when loading the "actions.bombard_max_range"
entry would fall through in error situations to error handling in 'default'.
This would not work for ENTRY_INT, as it would end up in ENTRY_STR and not
directly to 'default'.

Rewrite it with if, else if and else.

Reported by Marko Lindqvist

See hrm Bug #853675

(cherry picked from commit 00fb6766f9f9bafeba249762d8249ab37773599e)