Server: change all specialists, make specialist AND assign to type
This is the server implementation for the commit before.
Shift-clicking will now change all specialists with one packet.
Unassigned city workers can now default to a different type of specialist if it is encded in the highest 3 bits of the city_id (NB: limits total number of cities to about 8000 instead of 65000). Should be fine. This could be reworked with a new packet parameter and addition to the data structure but would break backward compatibility.