unittools.c: Fix bug in UWT/delay GOTO. 2 versions, read more:
unittools.c - this is the simple less risky fix that simply inserts a unit_forget_last_activity() before issuing a delayed GOTO, so that the user can't change the order before the GOTO executes and then "inherit" a wait from the delayed GOTO that will execute this turn. it works because of the convenient fact that resets changed_from and that's somewhat coincidentally all that's needed to patch the bug, but is potentially not accomodating to future structural changes or inner mechanics.

unittools.c.same_turn_waits.bak - (currently in private local repo only) this is a more elaborate fix which is "better" but vetoed because more radical restructuring is riskier than a single one line fix, but it should be saved in case needed later. Instead of ASSUMING all orders come from before TC and get fulfilled AFTER TC, (which isn't true because of same-turn GOTOs delayed by UWT), it:
1. always wipes existing waits always before making a new one
2. has logic to handle whether the wait is introduced on the same turn or is inherited from last turn
3. logic for how UWT-delayed goto that can be copied/modelled to generalise to other future actions that also gt delayed and fulfilled on the same turn
4. unit_wait_list_add is a front for unit_wait_list_append and cleans old waits first
5. unit_wait_list_wipe is a front for unit_wait_list_erase and clears the punit's record of waits that will be fulfilled on the current turn, from the punit->server.wait field.
The above allows waits to be introduced at any time which will be fulfilled on the current turn or next turn, and for them to both co-exist rather easily.