Change map Layer order : ...
<test and likely new standard for FCW layer ordering>

Before: Rivers and Specials were going over roads/rails. This was cleaner in a lot of cases, and hid the specials less. But in some cases like Forest River, Peat River, etc., it made roads hard to see and rails nearly impossible to see, resulting in the need to constantly info-click the tile.

After:  Roads/Rails go over rivers/specials. In many cases this will be more 'busy' and give a less clean/pretty look at Specials, which will still be visible however. However, rivers and specials won't hide a road or rail, eliminating the need for info-clicks. This is believed to maybe be a better trade-off, hence the experiment.