Test 2, re-ordered layers....
1. remove LAYER_ROADS totally, better efficiency in iterating tiles for each layer
2. make roads/rails a sublayer of LAYER_SPECIAL1

This is the only way to allow a special-over-rail-over-river sequencing.  The last commit tried rail-over-special-over-river, which was better than the previous special-over-river-over-rail, but also showed the reason why someone chose order, because it makes the roads/rails a bit messy. In theory, this solution of making roads/rails a sublayer inside LAYER_SPECIAL1, now allows the best of both worlds with the special-over-rail-over-river ordering and a performance gain.