Bring in mapimg fix
In freeciv.org side the commit fixed division by zero error, and
badly calculated player color width (with a high number of players).
FCW had a local modification in place. I think that was an attempt
to fix the division by zero error. It did nothing to the bad
calculation.
I decided to simply discard the old FCW modification, and took
the freeciv.org code in as is.
There still was a patch in the old patches directory to make
the FCW modification. I removed that for making things a tiny
bit more consistent now that such a modification is not in place.

hrm Bug #707912
Freeciv.org commit id: b098ab7bcd1354c86a585fffa2de70abcb0d12f3

Signed-off-by: Marko Lindqvist <cazfi74@gmail.com>