Bring in NeverProtects fix
Apparently it has not been the final freeciv.org NeverProtects
patch that has been applied to fcw.org. This updates it to one
that was applied to freeciv.org as
d567a724498ec1763168009c637d8b0759c2b367
from hrm Feature #809471.

Comment about the change in the final version from the ticket:

- Make sure there's no other reasons attack is not permitted than
  unreachability, before allowing it

Signed-off-by: Marko Lindqvist <cazfi74@gmail.com>