Temporary mp2-brava2 ruleset !!!NB!!!....
NOT compliant with new server updates,
NOT compatible with new server.

This ruleset should be deleted when possible. It is a temporary compatibility patch that only works on current production server as of the date of this commit, to give mp2-brava the ability to load scenario maps. After new mp2-brava for new server version is up and running, this ruleset should be DELETED.