/alias command covers more cases
Was set to only work with user_turns==1, but it turns out longturn games (and possibly other cases) start with user_turns==0, unlike other cases where it appears to start with user_turns==1. Due to this, the requirement is more liberalized and it will function if !(userturns>2) i.e. user_turns<=2.