Tax rate granularity; fix Hamilton rounding in taxes.
Before, a 50/50 split between bulbs and gold could literally yield a result of 26 bulbs and 3 gold. This is from using index-prioritised Hamilton method; which is now upgraded with a more modern and improved implementation of Hamilton's method. In addition, if Hamilton method is so clunky/sensitive to yield such skews, bring in possibility for rulesets to use finer granularity, which in some cases can be used to tune out skewing.