Top 5 cities more accurate formula for "Greatness"
Was size + great_wonders*5, which was highly inaccurate.

Now accounts for shield, gold, science output according to carefully calculated proportions; wonders AND other buildings; and slight adjustment for happiness/unhappiness.

Net effect associates a substantial reality to the "greatness" and a bit of mystery to how the city got its greatness.