Enhance flexibility of ruleset revolution mechanics
1. Introduce static bool immediate_revolution(pplayer) which returns true IFF a player can immediately transition to a new gov.
1a. Introduce the (hacky) sample feature in that function which expands ruleset control over this dynamic. Currently by hacking EFT_NO_ANARCHY to not allow immediate transition if EFT_NO_ANARCHY >= 10, but later can be a different EFT, game server settings, etc. The idea here is that logic for future expansions become centralized here.
2. Introduce static struct government *revolution_government() which centralizes all game logic for what government a player has to go through during a revolution (if any).
3. revolution_length, which is thankfully already centralized, takes minor changes to tie into the new stuff above.
4. handle_player_change_government(), government_change(), update_revolution() altered to tie into all the above.

NOTES: As rulesets start exploring greater diversity of government, such as mixture of stronger bonuses and penalties, the previous behaviors of (A) instantly changing government as often as you like after a revolution and (B) instantly changing government as often as you like in one turn; these created OP "SuperGov" conditions where at will, and especially with Statue of Liberty, one could in real-time avoid whatever penalties they needed to for a moment, gain whatever bonuses suited them for a moment.
The proof of concept (and fully working) hacky fix to this for EFT_NO_ANARCHY >= 10 (see above), fixes the "SuperGov" OP imbalance rulesets with greater diversity in governments, and can be replaced at any time in isolated centralized functions with new EFFECTS and/or game settings.