Revert "Make broadcast_city_info() fix identical to upstream master"
this reverts a change that was to make FCW code identical to upstream, which was a change to an earlier commit that was not broadcasting to the correct pplayer but to the pcity->owner instead.  The re-ordering of the if's by cazfi was thought to be a performance enhancement but turns out to make a fall-through crack in the "outer" else which is intended to catch the condition of !can_player_see_city_internals, but no longer caught that due to the re-ordering of those if's.