Fix glitch in double-click add to worklist followed by single-click then "Add to Worklist" button
Double clicking to add several improvements to worklist (for some reason, not units), then single clicking a new selection and manually clicking "add to worklist" would not only add that single item to the worklist but the history of everything you double clicked before. This makes sure the production_selection[] array is cleared after a double-click to add to worklist.