REVERT CANDIDATE: tile claiming base mechanics:
1. When a base claims a tile, make it actually do a ptile->claimer to really claim it.
2. Therefore when the base-extra is destroyed , it doesn't illogically forget
whose tile it is being worked. This allows extra such as tile_claim to claim a
tile and then not need the extra to continue to be there; or for an ally to fortress a tile agreed to be given and later after the city is working it, to pillage the fortress; and other things.
3. Make a destroyed tile extra maintain current ownership for 1 turn so that TC ownership calculations don't make foolish glitchy mistakes that don't have a chance to be reacted to by players.
The ownership change of tiles comes