CRF_GRANULAR_COMBAT_STRENGTH ...
Allow rulesets to use 10x values for combat strength to achieve granularity, but since 95% of units will be integer when divided by 10, we let each ruleset set a divisor to put these numbers in a more comfortable familiar range, such as Warriors/Armor attacking at 1 and 10 instead of 10 and 100. Something attacking at 48, however, will show everywhere in the UI as 4.8 and not 48, for example, if the ruleset sets the divisor of CRF_GRANULAR_COMBAT_STRENGTH to 10. (Can be set to whatever btw.)