HACK FIX error caused by upstream 0a92b6869180ae94dd23840f9ed36716bc1431d9
cityturn.c::worklist_change_build_target() had checks for improvement req removed, resulting in a "false postponement" where it was reported postponed but wasn't and became the actual prod target even though not allowed, leading to exploits. The HACK-PATCH-FIX is a "return false" in worklist_item_postpone_req_vec which is ironic because FALSE means to cancel and not postpone, but because of some other bug in the commit id (below), actually makes it postpone, which was the desired behaviour we wanted. blamed:0a92b6869180ae94dd23840f9ed36716bc1431d9