MP2D,E Landroad embark bug-fix
Give LandRoad class a fall-thru bug fix like the rest of Land units, to fix the server's flawed use of "move to tile then Board" as a proxy for "Transport Embark". Said bug neglects that after it moves to the tile and subtracts move points, the changed number of move points may make Boarding no longer legal within that ruleset.
The REAL FIX for this will be to use a temp var to temporarily add the move points back right before doing the embark actionenabler legality action, then after success, going back to a saved var representing the appropriate number of movepoints.