MP2D,E Prepare for Deboard to unsupported tile to emulate "disembark-to-same-tile"
Allows the logic that if you can disembark to an adjacent tile, you should at least be able to disembark to same tile for same move cost as a disembark, especially in those cases where deboarding to that tile isn't legal.
Prepares for "liberalizing" client and server to have less hard-coded rules and allow ruleset flexibility in deboard legalities as well.