Fix units with (moves_left ≤ 1) can't EMBARK
Fix bug where unit executes embark action (which is a really a pseudo-action that executes 2 actions of  order_move then order_board.) The problem was that order_move can deplete moves_left which changes the legality of order_board.

Also it assumed order_board will always be legal on the transport's tile in all cases where embark was legal from the adjacent tile (conceivably not the case in some theoretic rulesets).