MP2D,E: Allow Cargo with moves_left to embark to anywhere
The global restriction against embarking on LandRoad and LandRail exposed a server problem. GOTO can't test for actionenabler TRANSPORT_EMBARK without a huge performance hit so uses simpler functions as a proxy for that. As a result, GOTO shows as legal when the proxy says embark is possible (but it's really not). In the case of Cargo units, which aren't able to survive/exist on most tiles, this can "strand them". For this reason and for reduced micromanagement, simply make an exception that Cargo unit_class can always embark.