Prioritize display of famine feared over lawlessness
The lawless sprite over a city mostly hid the "empty plate" for famine feared. Thus if there is famine feared it's a given it will be lawless next turn so make sure that's more visible in such case. For "lawless feared" a.k.a. disorder, put that on top of the plate since they meld nicely.