Client: hard-coded MP2E boarding rules for Cargo Plane which has restricted subclasses of passengers
TODO: Need some other way to specify cargo classes for a unit_type instead of unit_class. This might aggregate to unit_classes which are best in most cases.
Suggest transport_allowed = "Type", "Type" as separate field in unit_type.
This will be a big step toward needing to hard-coded boarding rules on the client side.