IMPORTANT NOTES PLEASE READ. Fix bombard vigil's getting overridden by overly stickler pessimist odds ranges: ...
An actual fix/change in code is below, but first an IMPORTANT NOTE: OLD UPSTREAM code had:
'p_aa_unit' can attack 'p_moving_unit' but it may be not reciprocal. */
odds_moving_unit_wins = 1.0;   (var's upstream name was punitwin or something like that)

meaning if the moving unit's odds to attack the candidate autoattacker are 0 because it got a NULL when trying to find a defender on the autoattacker's tile (i.e. autoattacker is unreachable to it), then all this meant its odds to win were set to 1.0. When instead if it's unreachable you'd think it's set to 0.0. Theoretically the result of this would be elevated odds of moving unit to win and thus more encouraging the vigil unit to attack, when the reverse is true and its unreachability gives it a smugger position to not take bad odds on the attack.
Changing a 1.0 to a 0.0 in code that was running all these years is scary so this needs to be watched and blamed if anything pops up that isn't right.

1. PREVIOUS attempt to override for bombard-vigil units was unknowingly never implemented because the recalculation of the odds later, somehow reset the overrides from can_bombard (and harmless_cargo) and in spite of logic to unreset/ignore that, it slipped through. So now it also resets to prob.min and prob.max in spite of a prob.min of 0 was getting an override later. Now it's triple set to prevent this.