Bring in UTYF_ONLY_HITS_TARGETS
Note: previous commit overlaps this one with some code in unittools:kill_unit().

An anti-ballistic missile intercepts a cruise missile over a tile with 30 spies. It succeeds in killing it. Should 30 spies on the ground die? No. And what if it's v3 musketeers on the ground, on a mountain-river tile. Do they have some chance to stop the missile? No, of course not. Or it shouldn't be. Well, this UTYF fixes that and also another situation where we assume reachability for all units and then make some unreachable and list them as targets. This doesn't allow us to make ATTACKERS unable to reach certain types. There is no nontargets = "" list to override reachability. This would cause a sophisticated ruleset to have to define EVERY unit class as unreachable and hard code in every target. 😱. Or we could just add this UTYF and then you can only reach units in your targets list. Done and done!

Mind you, because this use case was for missiles, there is a side effect of only killing one unit and not [conventional] all reachable targets in the stack.

👉🏽 If we want to differentiate that we can override with an EFT another utype flag, or pull it from extra_bombard_stats for number of units killable in a stack, setting it between 1 and whatever.

The only thing needed to do that is putting one little && !(xyz) in that one line of code in kill_unit() and we can have it both ways. 👀