Experiment: Small bottom margin may prevent clipping in some cases for displaying Foreigners in city
Note this happened on 1685x925 screen as reported by MikeDuk but I was not able to replicate, however, I did see only narrowest of all space in the panel (0px) so that any slight overreach would indeed cause clipping, which I simulated other ways. Then adding this margin fixed the simulation of the clipping.